+++ pom.xml
@@ -8,7 +8,7 @@
 		<version>2.7.8</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.elbia</groupId>
+	<groupId>com.setokaio</groupId>
 	<artifactId>employee-management</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>employee-management</name>
@@ -47,6 +47,11 @@
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-validation</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.springdoc</groupId>
+			<artifactId>springdoc-openapi-ui</artifactId>
+			<version>1.6.2</version>
 		</dependency>
 	</dependencies>
