--- src/main/java/com/elbia/employeemanagement/Application.java
@@ -1,13 +1,0 @@
-package com.elbia.employeemanagement;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Application.class, args);
-	}
-
-}
