--- src/main/java/com/elbia/employeemanagement/domain/constants/EmployeeStatus.java
@@ -1,5 +1,0 @@
-package com.elbia.employeemanagement.domain.constants;
-
-public enum EmployeeStatus {
-    ATIVO, DEMITIDO, REJEITADO_POR_FALTA_DE_RECURSOS
-}
