--- src/main/java/com/elbia/employeemanagement/domain/repository/DepartamentoRepository.java
@@ -1,16 +1,0 @@
-package com.elbia.employeemanagement.domain.repository;
-
-import com.elbia.employeemanagement.domain.entity.Departamento;
-import com.elbia.employeemanagement.domain.entity.Funcionario;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import javax.transaction.Transactional;
-import java.util.Optional;
-
-@Repository
-@Transactional
-public interface DepartamentoRepository extends JpaRepository<Departamento, Long> {
-
-
-}
