--- src/main/java/com/elbia/employeemanagement/domain/repository/FuncionarioRepository.java
@@ -1,15 +1,0 @@
-package com.elbia.employeemanagement.domain.repository;
-
-import com.elbia.employeemanagement.domain.entity.Funcionario;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import javax.transaction.Transactional;
-import java.util.Optional;
-
-@Repository
-@Transactional
-public interface FuncionarioRepository extends JpaRepository<Funcionario, Long> {
-
-    public Optional<Funcionario> findByCpf(String cpf);
-}
