--- src/main/java/com/elbia/employeemanagement/domain/utils/Validation.java
@@ -1,15 +1,0 @@
-package com.elbia.employeemanagement.domain.utils;
-
-import com.elbia.employeemanagement.domain.entity.Departamento;
-
-import java.math.BigDecimal;
-
-public final class Validation {
-    private Validation(){}
-
-    public static boolean validaOrcamento(Departamento departamento, BigDecimal valor){
-        BigDecimal totalDeSalarios = departamento.getTotalSalarios();
-        int comparacao = totalDeSalarios.add(valor).compareTo(departamento.getBudget());
-        return comparacao < 1;
-    }
-}
