+++ src/main/java/com/setokaio/employeemanagement/Application.java
@@ -1,0 +1,13 @@
+package com.setokaio.employeemanagement;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+
+	public static void main(String[] args) {
+		SpringApplication.run(Application.class, args);
+	}
+
+}
