+++ src/main/java/com/setokaio/employeemanagement/domain/constants/EmployeeStatus.java
@@ -1,0 +1,5 @@
+package com.setokaio.employeemanagement.domain.constants;
+
+public enum EmployeeStatus {
+    ATIVO, DEMITIDO, REJEITADO_POR_FALTA_DE_RECURSOS
+}
