+++ src/main/java/com/setokaio/employeemanagement/domain/repository/FuncionarioRepository.java
@@ -1,0 +1,15 @@
+package com.setokaio.employeemanagement.domain.repository;
+
+import com.setokaio.employeemanagement.domain.entity.Funcionario;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import javax.transaction.Transactional;
+import java.util.Optional;
+
+@Repository
+@Transactional
+public interface FuncionarioRepository extends JpaRepository<Funcionario, Long> {
+
+    Optional<Funcionario> findByCpf(String cpf);
+}
