--- src/test/java/com/elbia/employeemanagement/ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.elbia.employeemanagement;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
