+++ src/test/java/com/setokaio/employeemanagement/ApplicationTests.java
@@ -1,0 +1,13 @@
+package com.setokaio.employeemanagement;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
